Refactor code and create Action class
