+++ src/main/java/Deadline.java
@@ -1,0 +1,12 @@
+public class Deadline extends Task {
+    protected String by;
+    public Deadline(String description, String by) {
+        super(description);
+        this.by =by;
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + by + ")";
+    }
+}
