+++ src/main/java/ElonException.java
@@ -1,0 +1,5 @@
+public class ElonException extends Exception{
+    ElonException(String message) {
+        super(message);
+    }
+}
